==> notebooks/1-preprocessing.ipynb <==
import pandas as pd
import numpy as np
import transformers
import nltk
import spacy
import datasets
from transformers import BertTokenizer, BertModel
from datasets import load_dataset
from datasets import Dataset, DatasetDict
from sklearn.feature_extraction.text import TfidfVectorizer

ds = load_dataset('multi_news')

ds

ds['train'][0]

ds['train'].features

ds_train = pd.DataFrame(ds['train'])
ds_test = pd.DataFrame(ds['test'])
ds_validation = pd.DataFrame(ds['validation'])

# remove punctuation, capitalization and stop words from ds_train, ds_test, ds_validation in place
nltk.download('stopwords')
from nltk.corpus import stopwords

stop_words = set(stopwords.words('english'))

ds_train['document'] = ds_train['document'].apply(lambda x: ' '.join([word for word in x.split() if word not in stop_words]))
ds_test['document'] = ds_test['document'].apply(lambda x: ' '.join([word for word in x.split() if word not in stop_words]))
ds_validation['document'] = ds_validation['document'].apply(lambda x: ' '.join([word for word in x.split() if word not in stop_words]))


# convert all uppercase to lower case
ds_train['document'] = ds_train['document'].str.lower()
ds_test['document'] = ds_test['document'].str.lower()

# remove special characters 
ds_train['document'] = ds_train['document'].str.replace('[^a-zA-Z0-9\s]','', regex=True)
ds_test['document'] = ds_test['document'].str.replace('[^a-zA-Z0-9\s]','', regex=True)

#remove string "(ap)"  from all rows in ds_train and ds_test
ds_train['document'] = ds_train['document'].str.replace(r'\(ap\)','', regex=True).fillna('')
ds_test['document'] = ds_test['document'].str.replace(r'\(ap\)','', regex=True).fillna('')



ds_train['document']

# Convert dataframes into Dataset objects
# assign the splits
train = Dataset.from_pandas(ds_train)
test = Dataset.from_pandas(ds_test)

# reconstruct both datasets into a Dataset Dict object
new_ds = DatasetDict(
    {
        'train': train,
        'test': test
    }
)

new_ds

new_ds['train']

new_ds['train']['document'][0]

# Convert DatasetDict to DataFrames
train_df = pd.DataFrame(new_ds['train'])
test_df = pd.DataFrame(new_ds['test'])

# Export DataFrames to CSV
train_df.to_csv('../data/processed/new_ds_train_dataset.csv', index=False)
test_df.to_csv('../data/processed/new_ds_test_dataset.csv', index=False)


ds_validation.to_csv('../data/validation_set.csv', index=False)
