==> notebooks/5-deployment.ipynb <==
from huggingface_hub import notebook_login
notebook_login()

 !pip install git-lfs

from transformers import AutoTokenizer, Seq2SeqTrainer, Seq2SeqTrainingArguments, AutoModelForSeq2SeqLM, DataCollatorForSeq2Seq, AutoModelForSequenceClassification, TrainingArguments, Trainer, Trainer, TrainingArguments
from datasets import Dataset, DatasetDict
from datasets import load_dataset
import evaluate
import pandas as pd
import nltk
import numpy as np

nltk.download('punkt')
import warnings
# Suppress all warnings
warnings.filterwarnings('ignore')
warnings.filterwarnings("ignore", category=UserWarning, module='transformers')
warnings.filterwarnings("ignore", category=UserWarning)


# Load data
test_dataset = pd.read_csv('../data/processed/new_ds_test_dataset.csv')
train_dataset = pd.read_csv('../data/processed/new_ds_train_dataset.csv')
validation_dataset = pd.read_csv('../data/validation_set.csv')

# Ensure all documents are strings
train_dataset['document'] = train_dataset['document'][:150].astype(str)
train_dataset['summary'] = train_dataset['summary'][:150].astype(str)
test_dataset['document'] = test_dataset['document'][:50].astype(str)
test_dataset['summary'] = test_dataset['summary'][:50].astype(str)
validation_dataset['document'] = validation_dataset['document'][:50].astype(str)
validation_dataset['summary'] = validation_dataset['summary'][:50].astype(str)

# Handle missing values (remove rows with missing documents or summaries)
train_dataset.dropna(subset=['document', 'summary'], inplace=True)
test_dataset.dropna(subset=['document', 'summary'], inplace=True)
validation_dataset.dropna(subset=['document', 'summary'], inplace=True)

# Convert the pandas DataFrames to Hugging Face Datasets
train = Dataset.from_pandas(train_dataset)
test = Dataset.from_pandas(test_dataset)
validation = Dataset.from_pandas(validation_dataset)

# Reconstruct both datasets into a Dataset Dict object
new_ds = DatasetDict({
    'train': train,
    'test': test,
    'validation': validation
})

tokenizer = AutoTokenizer.from_pretrained("sshleifer/distilbart-cnn-12-6")

prefix = "summarize: "

def preprocess_function(examples):
    inputs = [prefix + doc for doc in examples["document"]]
    model_inputs = tokenizer(inputs, max_length=1024, truncation=True)

    labels = tokenizer(text_target=examples["summary"], max_length=128, truncation=True)

    model_inputs["labels"] = labels["input_ids"]
    return model_inputs

tokenized_datasets = new_ds.map(preprocess_function, batched=True)

# Setup evaluation
nltk.download("punkt", quiet=True)
metric = evaluate.load("rouge")


def compute_metrics(eval_preds):
    preds, labels = eval_preds

    # decode preds and labels
    labels = np.where(labels != -100, labels, tokenizer.pad_token_id)
    decoded_preds = tokenizer.batch_decode(preds, skip_special_tokens=True)
    decoded_labels = tokenizer.batch_decode(labels, skip_special_tokens=True)

    # rougeLSum expects newline after each sentence
    decoded_preds = ["\n".join(nltk.sent_tokenize(pred.strip())) for pred in decoded_preds]
    decoded_labels = ["\n".join(nltk.sent_tokenize(label.strip())) for label in decoded_labels]

    result = metric.compute(predictions=decoded_preds, references=decoded_labels, use_stemmer=True)
    return result

# Load pretrained model and evaluate model after each epoch
model = AutoModelForSeq2SeqLM.from_pretrained("sshleifer/distilbart-cnn-12-6")
data_collator = DataCollatorForSeq2Seq(tokenizer=tokenizer, model=model)

#### Optimization from previous notebook - also pushing this to repo on hugging-face

###### Changes

**learning_rate:** Increased from `2e-5` to `5e-5` to help the model converge faster.

**per_device_train_batch_size:** Increased from `16` to `32` to allow the model to learn from more examples in each step.

**per_device_eval_batch_size:** Increased from `4` to `8` to speed up evaluation.

**weight_decay:** Increased from `0.01` to `0.05` for better regularization.

**num_train_epochs:** Increased from `2` to `5` to give the model more time to learn.

**lr_scheduler_type:** Added `linear` learning rate scheduler to adjust the learning rate dynamically during training.

repo_name = "imarri01-huggingface"
training_args = Seq2SeqTrainingArguments(
    output_dir=repo_name,
    evaluation_strategy="epoch",
    learning_rate=5e-5,  # Increased learning rate
    per_device_train_batch_size=32,  # Increased train batch size
    per_device_eval_batch_size=8,  # Increased eval batch size
    weight_decay=0.05,  # Increased weight decay
    save_total_limit=3,
    num_train_epochs=5,  # Increased number of epochs
    fp16=True,
    predict_with_generate=True,
    lr_scheduler_type="linear",  # Added learning rate scheduler
    push_to_hub=True,
)

trainer = Seq2SeqTrainer(
    model=model,
    args=training_args,
    train_dataset=tokenized_datasets["train"],
    eval_dataset=tokenized_datasets["test"],
    tokenizer=tokenizer,
    data_collator=data_collator,
    compute_metrics=compute_metrics
)

trainer.evaluate()

The optimization improved the model run time from 1hr down to 6 mins

trainer.push_to_hub()

URL: https://huggingface.co/imarri01/imarri01-huggingface
